# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: ComIT Viribus API
  description: REST API для SPA (без эндпоинтов ИИ-чата)
  version: 1.0.0
paths:
...
  /api/news/mini:
    get:
      tags:
      - news
      summary: Get News Mini
      operationId: get_news_mini_api_news_mini_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/NewsMiniItem'
                type: array
                title: Response Get News Mini Api News Mini Get
    post:
      tags:
      - news
      summary: Create News Mini
      operationId: create_news_mini_api_news_mini_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewsMiniCreateBody'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewsMiniItem'
        '409':
          description: Уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /api/news/mini/{news_id}:
    patch:
      tags:
      - news
      summary: Patch News Mini
      operationId: patch_news_mini_api_news_mini__news_id__patch
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: news_id
        in: path
        required: true
        schema:
          type: string
          title: News Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewsMiniUpdateBody'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NewsMiniItem'
        '404':
          description: Не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '403':
          description: Нет прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - news
      summary: Delete News Mini
      operationId: delete_news_mini_api_news_mini__news_id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: news_id
        in: path
        required: true
        schema:
          type: string
          title: News Id
      responses:
        '204':
          description: Successful Response
        '404':
          description: Не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '403':
          description: Нет прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/news/featured:
    get:
      tags:
      - news
      summary: Get News Featured
      operationId: get_news_featured_api_news_featured_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/FeaturedNewsItem'
                type: array
                title: Response Get News Featured Api News Featured Get
    post:
      tags:
      - news
      summary: Create News Featured
      operationId: create_news_featured_api_news_featured_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewsFeaturedCreateBody'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeaturedNewsItem'
        '409':
          description: Уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '403':
          description: Только модератор
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /api/news/featured/{news_id}:
    patch:
      tags:
      - news
      summary: Patch News Featured
      operationId: patch_news_featured_api_news_featured__news_id__patch
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: news_id
        in: path
        required: true
        schema:
          type: string
          title: News Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewsFeaturedUpdateBody'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeaturedNewsItem'
        '404':
          description: Не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '403':
          description: Только модератор
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - news
      summary: Delete News Featured
      operationId: delete_news_featured_api_news_featured__news_id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: news_id
        in: path
        required: true
        schema:
          type: string
          title: News Id
      responses:
        '204':
          description: Successful Response
        '404':
          description: Не найдено
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '403':
          description: Только модератор
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/projects:
    post:
      tags:
      - projects
      summary: Create Project
      operationId: create_project_api_projects_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreateBody'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDetails'
        '404':
          description: Колонка не найдена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '409':
          description: Проект с таким id уже есть
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
...
  /api/projects/{project_id}:
    patch:
      tags:
      - projects
      summary: Patch Project
      operationId: patch_project_api_projects__project_id__patch
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdateBody'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDetails'
        '404':
          description: Проект не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '403':
          description: Нет прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - projects
      summary: Delete Project
      operationId: delete_project_api_projects__project_id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      responses:
        '204':
          description: Successful Response
        '404':
          description: Проект не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '403':
          description: Нет прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    get:
      tags:
      - projects
      summary: Get Project By Id
      operationId: get_project_by_id_api_projects__project_id__get
      parameters:
      - name: project_id
        in: path
        required: true
        schema:
          type: string
          title: Project Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectDetails'
        '404':
          description: Проект не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/admin/projects/columns:
    post:
      tags:
      - projects-admin
      summary: Admin Create Column
      operationId: admin_create_column_api_admin_projects_columns_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectsColumnCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectColumn'
        '409':
          description: Колонка уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - OAuth2PasswordBearer: []
  /api/admin/projects/columns/{column_id}:
    patch:
      tags:
      - projects-admin
      summary: Admin Update Column
      operationId: admin_update_column_api_admin_projects_columns__column_id__patch
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: column_id
        in: path
        required: true
        schema:
          type: string
          title: Column Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectsColumnUpdate'
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    delete:
      tags:
      - projects-admin
      summary: Admin Delete Column
      operationId: admin_delete_column_api_admin_projects_columns__column_id__delete
      security:
      - OAuth2PasswordBearer: []
      parameters:
      - name: column_id
        in: path
        required: true
        schema:
          type: string
          title: Column Id
      responses:
        '204':
          description: Successful Response
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ErrorDetail:
      properties:
        detail:
          type: string
          title: Detail
      type: object
      required:
      - detail
      title: ErrorDetail
      description: Тело ответа HTTPException (например 404).
...
    FeaturedNewsItem:
      properties:
        id:
          type: string
          title: Id
        title:
          type: string
          title: Title
        subtitle:
          type: string
          title: Subtitle
        description:
          type: string
          title: Description
        imageUrl:
          type: string
          title: Imageurl
        ctaLabel:
          type: string
          title: Ctalabel
        detailsUrl:
          type: string
          title: Detailsurl
      type: object
      required:
      - id
      - title
      - subtitle
      - description
      - imageUrl
      - ctaLabel
      - detailsUrl
      title: FeaturedNewsItem
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    NewsFeaturedCreateBody:
      properties:
        id:
          type: string
          title: Id
        title:
          type: string
          title: Title
        subtitle:
          type: string
          title: Subtitle
        description:
          type: string
          title: Description
        imageUrl:
          type: string
          title: Imageurl
        ctaLabel:
          type: string
          title: Ctalabel
        detailsUrl:
          type: string
          title: Detailsurl
      type: object
      required:
      - id
      - title
      - subtitle
      - description
      - imageUrl
      - ctaLabel
      - detailsUrl
      title: NewsFeaturedCreateBody
    NewsFeaturedUpdateBody:
      properties:
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        subtitle:
          anyOf:
          - type: string
          - type: 'null'
          title: Subtitle
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        imageUrl:
          anyOf:
          - type: string
          - type: 'null'
          title: Imageurl
        ctaLabel:
          anyOf:
          - type: string
          - type: 'null'
          title: Ctalabel
        detailsUrl:
          anyOf:
          - type: string
          - type: 'null'
          title: Detailsurl
      type: object
      title: NewsFeaturedUpdateBody
    NewsMiniCreateBody:
      properties:
        id:
          type: string
          title: Id
        title:
          type: string
          title: Title
        imageUrl:
          type: string
          title: Imageurl
        detailsUrl:
          type: string
          title: Detailsurl
      type: object
      required:
      - id
      - title
      - imageUrl
      - detailsUrl
      title: NewsMiniCreateBody
    NewsMiniItem:
      properties:
        id:
          type: string
          title: Id
        title:
          type: string
          title: Title
        imageUrl:
          type: string
          title: Imageurl
        detailsUrl:
          type: string
          title: Detailsurl
      type: object
      required:
      - id
      - title
      - imageUrl
      - detailsUrl
      title: NewsMiniItem
    NewsMiniUpdateBody:
      properties:
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        imageUrl:
          anyOf:
          - type: string
          - type: 'null'
          title: Imageurl
        detailsUrl:
          anyOf:
          - type: string
          - type: 'null'
          title: Detailsurl
      type: object
      title: NewsMiniUpdateBody
...
    PersonPreview:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        avatarUrl:
          type: string
          title: Avatarurl
        avatarVariant:
          type: string
          title: Avatarvariant
      type: object
      required:
      - id
      - name
      - avatarUrl
      - avatarVariant
      title: PersonPreview
    ProductivityBlock:
      properties:
        value:
          type: string
          title: Value
        delta:
          type: string
          title: Delta
      type: object
      required:
      - value
      - delta
      title: ProductivityBlock
...
    ProgressBlock:
      properties:
        value:
          type: string
          title: Value
        percent:
          type: integer
          title: Percent
      type: object
      required:
      - value
      - percent
      title: ProgressBlock
    ProjectColumn:
      properties:
        id:
          type: string
          title: Id
        title:
          type: string
          title: Title
        count:
          type: integer
          title: Count
        projects:
          items:
            $ref: '#/components/schemas/HubProject'
          type: array
          title: Projects
      type: object
      required:
      - id
      - title
      - count
      - projects
      title: ProjectColumn
    ProjectCreateBody:
      properties:
        id:
          type: string
          title: Id
        columnId:
          type: string
          title: Columnid
        code:
          type: string
          title: Code
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
        teamName:
          type: string
          title: Teamname
        updatedLabel:
          type: string
          title: Updatedlabel
        teamAvatarUrl:
          type: string
          title: Teamavatarurl
        detailsUrl:
          type: string
          title: Detailsurl
        visibility:
          anyOf:
          - type: string
            enum:
            - public
            - private
          - type: 'null'
          title: Visibility
        isHot:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Ishot
        ownerName:
          type: string
          title: Ownername
        joinLabel:
          type: string
          title: Joinlabel
        teamCaption:
          type: string
          title: Teamcaption
        productivityCaption:
          type: string
          title: Productivitycaption
        progressCaption:
          type: string
          title: Progresscaption
        todoCaption:
          type: string
          title: Todocaption
        integrationCaption:
          type: string
          title: Integrationcaption
        teamMembersPreview:
          items:
            $ref: '#/components/schemas/PersonPreview'
          type: array
          title: Teammemberspreview
        productivity:
          $ref: '#/components/schemas/ProductivityBlock'
        progress:
          $ref: '#/components/schemas/ProgressBlock'
        todo:
          $ref: '#/components/schemas/TodoBlock'
        integrations:
          items:
            $ref: '#/components/schemas/IntegrationItem'
          type: array
          title: Integrations
        participants:
          items:
            $ref: '#/components/schemas/ParticipantRow'
          type: array
          title: Participants
      type: object
      required:
      - id
      - columnId
      - code
      - title
      - description
      - teamName
      - updatedLabel
      - teamAvatarUrl
      - detailsUrl
      - ownerName
      - joinLabel
      - teamCaption
      - productivityCaption
      - progressCaption
      - todoCaption
      - integrationCaption
      - teamMembersPreview
      - productivity
      - progress
      - todo
      - integrations
      - participants
      title: ProjectCreateBody
    ProjectDetails:
      properties:
        id:
          type: string
          title: Id
        code:
          type: string
          title: Code
        title:
          type: string
          title: Title
        ownerName:
          type: string
          title: Ownername
        detailsUrl:
          type: string
          title: Detailsurl
        joinLabel:
          type: string
          title: Joinlabel
        teamCaption:
          type: string
          title: Teamcaption
        productivityCaption:
          type: string
          title: Productivitycaption
        progressCaption:
          type: string
          title: Progresscaption
        todoCaption:
          type: string
          title: Todocaption
        integrationCaption:
          type: string
          title: Integrationcaption
        teamMembersPreview:
          items:
            $ref: '#/components/schemas/PersonPreview'
          type: array
          title: Teammemberspreview
        productivity:
          $ref: '#/components/schemas/ProductivityBlock'
        progress:
          $ref: '#/components/schemas/ProgressBlock'
        todo:
          $ref: '#/components/schemas/TodoBlock'
        integrations:
          items:
            $ref: '#/components/schemas/IntegrationItem'
          type: array
          title: Integrations
        participants:
          items:
            $ref: '#/components/schemas/ParticipantRow'
          type: array
          title: Participants
      type: object
      required:
      - id
      - code
      - title
      - ownerName
      - detailsUrl
      - joinLabel
      - teamCaption
      - productivityCaption
      - progressCaption
      - todoCaption
      - integrationCaption
      - teamMembersPreview
      - productivity
      - progress
      - todo
      - integrations
      - participants
      title: ProjectDetails
    ProjectUpdateBody:
      properties:
        columnId:
          anyOf:
          - type: string
          - type: 'null'
          title: Columnid
        code:
          anyOf:
          - type: string
          - type: 'null'
          title: Code
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        teamName:
          anyOf:
          - type: string
          - type: 'null'
          title: Teamname
        updatedLabel:
          anyOf:
          - type: string
          - type: 'null'
          title: Updatedlabel
        teamAvatarUrl:
          anyOf:
          - type: string
          - type: 'null'
          title: Teamavatarurl
        detailsUrl:
          anyOf:
          - type: string
          - type: 'null'
          title: Detailsurl
        visibility:
          anyOf:
          - type: string
            enum:
            - public
            - private
          - type: 'null'
          title: Visibility
        isHot:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Ishot
        ownerName:
          anyOf:
          - type: string
          - type: 'null'
          title: Ownername
        joinLabel:
          anyOf:
          - type: string
          - type: 'null'
          title: Joinlabel
        teamCaption:
          anyOf:
          - type: string
          - type: 'null'
          title: Teamcaption
        productivityCaption:
          anyOf:
          - type: string
          - type: 'null'
          title: Productivitycaption
        progressCaption:
          anyOf:
          - type: string
          - type: 'null'
          title: Progresscaption
        todoCaption:
          anyOf:
          - type: string
          - type: 'null'
          title: Todocaption
        integrationCaption:
          anyOf:
          - type: string
          - type: 'null'
          title: Integrationcaption
        teamMembersPreview:
          anyOf:
          - items:
              $ref: '#/components/schemas/PersonPreview'
            type: array
          - type: 'null'
          title: Teammemberspreview
        productivity:
          anyOf:
          - $ref: '#/components/schemas/ProductivityBlock'
          - type: 'null'
        progress:
          anyOf:
          - $ref: '#/components/schemas/ProgressBlock'
          - type: 'null'
        todo:
          anyOf:
          - $ref: '#/components/schemas/TodoBlock'
          - type: 'null'
        integrations:
          anyOf:
          - items:
              $ref: '#/components/schemas/IntegrationItem'
            type: array
          - type: 'null'
          title: Integrations
        participants:
          anyOf:
          - items:
              $ref: '#/components/schemas/ParticipantRow'
            type: array
          - type: 'null'
          title: Participants
      type: object
      title: ProjectUpdateBody
    ProjectsColumnCreate:
      properties:
        id:
          type: string
          title: Id
        title:
          type: string
          title: Title
      type: object
      required:
      - id
      - title
      title: ProjectsColumnCreate
    ProjectsColumnUpdate:
      properties:
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        sortOrder:
          anyOf:
          - type: integer
          - type: 'null'
          title: Sortorder
      type: object
      title: ProjectsColumnUpdate
...
    TodoBlock:
      properties:
        task:
          type: string
          title: Task
        updatedLabel:
          type: string
          title: Updatedlabel
      type: object
      required:
      - task
      - updatedLabel
      title: TodoBlock
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: /api/auth/jwt/login